requirements.txt
python-dotenv==1.0.0
openai==1.12.0
requests==2.31.0